added auth for create, delete and update

+++ app/api/DeleteLapById/route.tsx
@@ -1,6 +1,8 @@
 import prisma from 'components/prisma'
 import { NextRequest, NextResponse } from "next/server";
 import assert from 'assert';
+import { AVAILABLE_PERMISSIONS } from 'components/helpers/users';
+import { isRequestAuthorised } from 'components/helpers/auth';
 
 async function deleteLap(id: string) {
     var res = await prisma.lap.update({
@@ -20,15 +22,20 @@ export async function POST(request: NextRequest) {
         assert.notStrictEqual(undefined, req.id, 'Id required');
 
     } catch (bodyError) {
-        return NextResponse.json({ error: true, message: "information missing" });
+        return NextResponse.json({ error: "information missing" }, { status: 400 });
+    }
+
+    let authorised = await isRequestAuthorised(request.cookies.get("token")!.value, AVAILABLE_PERMISSIONS.editResults)
+    if (!authorised) {
+        return NextResponse.json({ error: "not authorized" }, { status: 401 });
     }
 
     var id = req.id
 
     var lap = await deleteLap(id)
-    if (!lap) {
-        return NextResponse.json({ error: true, message: 'Could not delete lap' });
+    if (lap) {
+        return NextResponse.json({ res: lap }, { status: 200 });
     }
-    return NextResponse.json({ error: false, result: lap });
+    return NextResponse.json({ error: 'Could not delete lap' }, { status: 400 });
 }
