+++ app/api/DeleteUserById/route.tsx
@@ -1,6 +1,8 @@
 import prisma from 'components/prisma'
 import { NextRequest, NextResponse } from "next/server";
 import assert from 'assert';
+import { AVAILABLE_PERMISSIONS } from 'components/helpers/users';
+import { isRequestAuthorised } from 'components/helpers/auth';
 
 async function findUser(user: UserDataType) {
     var result = await prisma.user.findFirst({
@@ -26,20 +28,21 @@ export async function POST(request: NextRequest) {
         assert.notStrictEqual(undefined, req.user, 'user required');
 
     } catch (bodyError) {
-        return NextResponse.json({ error: true, message: "information missing" });
+        return NextResponse.json({ error: "information missing" }, { status: 400 });
+    }
+
+    let authorised = await isRequestAuthorised(request.cookies.get("token")!.value, AVAILABLE_PERMISSIONS.editSeries)
+    if (!authorised) {
+        return NextResponse.json({ error: "not authorized" }, { status: 401 });
     }
 
     var user = req.user
-    console.log("this")
-    console.log(user)
 
     var deletedUser = await findUser(user)
     if (deletedUser) {
         await deleteUser(user)
-        return NextResponse.json({ error: false, user: deletedUser });
-    } else {
-        // User exists
-        return NextResponse.json({ error: true, message: 'user not found' });
+        return NextResponse.json({ res: deletedUser }, { status: 200 });
     }
+    return NextResponse.json({ error: 'user not found' }, { status: 400 });
 }
